crypto/requirements.txt
shrimpy-python
